+++ CHANGELOG.md
@@ -47,6 +47,8 @@ Be carefull when updating and check the changelog!
 - Identities of `TableReference` and `ColumnReference` objects are now based on their lowercase representations. This is a
   necessary change caused by the migration to pglast, since Postgres does not retain the original casing of identifiers after
   parsing.
+- Removed `BooleanExpression`. Instead, `AbstractPredicate` is a `SqlExpression` now. This removes the weird distinction
+  between predicates and expressions.
 - Removed illegal SQL operator "**MISSING**", which was an artifact of the mo-sql parser output
 - ``IS NULL`` and ``IS NOT NULL`` predicates are no longer represented by `UnaryPredicate` instances, but by `BinaryPredicate`
   with a *None* static value